// src/app/models/customer.model.ts
export class Customer {

     id?: string;
     salutation: string;
     initials: string;
     firstname: string;
     firstnameAscii: string;
     firstnameCountryRank: string;
     firstnameCountryFrequency: string;
     lastname: string;
     lastnameAscii: string;
     lastnameCountryRank: string;
     lastnameCountryFrequency: string;
     email: string;
     password: string;
     countryCode: string;
     countryCodeAlpha: string;
     countryName: string;
     primaryLanguageCode: string;
     primaryLanguage: string;
     balance: number;
     phoneNumber: string;
     currency: string;
     gender: string;
}